break github out into own file
